% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bs_basis}
\alias{bs_basis}
\title{Generate an evaluated basis matrix for B-splines}
\usage{
bs_basis(x, order, knots)
}
\arguments{
\item{x}{Predictor variable vector.}

\item{order}{The order of basis functions. order=degree+1}

\item{knots}{The knots used to construct the B-splines, including innerknots, boundary knots and phantom knots.
It can be generated by \code{bs_knots}.}
}
\value{
Basis matrix evaluated at each x value.
}
\description{
#' This function generates B-spline basis. The B-splines are defined following
the recursive formulas due to de Boor. Only univariate input can be used.
}
\examples{
x<-seq(0, 1, 0.001)
knots <- seq(0, 1, 0.1)
order<-4
knots<-bs_knots(x,real_knots=knots,order=order)

basis<-bs_basis(x,order,knots)
plot(x,rep(0,length(x)),type="l",ylim=c(0,1))
for (i in 1: (length(knots)+order)){
  lines(x,basis[,i])
}
}
