% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bs_knots}
\alias{bs_knots}
\title{Add phantom knots for B-splines}
\usage{
bs_knots(x, df, real_knots, q)
}
\arguments{
\item{x}{Predictor variable vector.}

\item{df}{Degrees of freedom. One can supply df rather than knots.}

\item{real_knots}{The innerknots and boundary knots that define the spline. The knots can all be innerknots.
The knots provided can be quantiles of x or real values.
More explanation of \code{knots}, \code{df}, \code{q} can be seen in \code{generate_knots}.}

\item{q}{A boolean variable define whether knots provided are quantiles or real values. When \code{q}=TRUE, \code{real_knots}
are quantiles of x. When \code{q}=FALSE, \code{real_knots} are real values of x. Default is FALSE.}

\item{order}{The order of basis functions. order=degree+1}
}
\value{
The knots used to construct the B-splines, including innerknots, boundary knots and phantom knots.
}
\description{
Add phantom knots for B-splines
}
