% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bs_train}
\alias{bs_train}
\title{Train regression coefficients for B-splines.}
\usage{
bs_train(x, y, order, real_knots = NULL, df = NULL, q = FALSE)
}
\arguments{
\item{x}{The input vector of training dataset.}

\item{y}{The output vector of training dataset.}

\item{order}{The order of B-spline functions. The default is order=4 for cubic B-splines.}

\item{real_knots}{The innerknots and boundary knots that define the spline.
Phantom knots should not be included. Phantom knots will be generated by \code{bs_knots}
The knots provided can be quantiles of x or real values.
More explanation of \code{knots}, \code{df}, \code{q} can be seen in \code{generate_knots}.}

\item{df}{Degrees of freedom. One can supply df rather than knots.}

\item{q}{A boolean variable define whether knots provided are quantiles or real values. When \code{q}=TRUE, \code{real_knots}
are quantiles of x. When \code{q}=FALSE, \code{real_knots} are real values of x. Default is FALSE.}
}
\value{
A list with the following components:
\item{beta}{The coefficients of nonparametric regression.}
\item{basis}{The B-spline basis matrix of dimension c(length(x), df). df = length(innerknots) + order.}
\item{knots}{The knots used to construct the B-splines, including innerknots, boundary knots and phantom knots}
\item{order}{The order of basis functions. order=degree+1}
}
\description{
Train regression coefficients for B-splines.
}
\examples{
x<-seq(0, 1, 0.001)
y <- x^3 * 3 - x^2 * 2 + x + exp(1)+rnorm(length(x),0,0.1)
plot(x,y)
knots <- seq(0, 1, 0.1)
order<-4

basis<-bs_train(x,y,order,knots)
plot(x,rep(0,length(x)),type="l",ylim=c(0,1))
for (i in 1: (length(knots)+order)){
  lines(x,basis$basismatrix[,i],col=i)
}
}
\references{
"Friedman, J., Hastie, T., & Tibshirani, R. (2001). The elements of statistical learning (Vol. 1, pp. 337-387). New York: Springer series in statistics,"
Chapter 5, Appendix.
}
\seealso{
\code{bs_knots}.
}
