% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bsplineFitting}
\alias{bsplineFitting}
\title{Regression using B-spline basis}
\usage{
bsplineFitting(x_test, basis)
}
\arguments{
\item{x_test}{The input values at which evaluations are required.}

\item{basis}{The return value of function \code{bsplinBasis}.}
}
\value{
The evaluated output at x_test.
}
\description{
This function provides nonparametric regressions using B-splines. The B-splines are generated by the function \code{bsplinBasis}.
The return value of \code{bsplinBasis} is required as an argument of \code{bsplineFitting}
}
\examples{
x<-seq(0, 1, 0.001)
y <- x^3 * 3 - x^2 * 2 + x + exp(1)+rnorm(length(x),0,0.1)
plot(x,y)
innerknots <- seq(0.1, 0.9, 0.01)
order<-4
basis<-bsplineBasis(x,y,order,innerknots)

x_test<-seq(0, 1, 0.01)
fit<-bsplineFitting(x_test,basis)
plot(x_test,fit)
lines(x_test,x_test^3 * 3 - x_test^2 * 2 + x_test + exp(1),col="red")
}
