% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{cal_loo_cv_error}
\alias{cal_loo_cv_error}
\title{Calculte leave-one-out CV error for a linear smoother}
\usage{
cal_loo_cv_error(y, f_hat, S)
}
\arguments{
\item{y}{response variable values}

\item{f_hat}{fitted response variable values}

\item{S}{smoother matrix}
}
\value{
leave-one-out cross-validation error
}
\description{
Calculte leave-one-out CV error for a linear smoother
}
