% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingspline.R
\name{css_predict}
\alias{css_predict}
\title{Prediction using smoothing spline with squared 2nd derivative penalty}
\usage{
css_predict(basis, x_test)
}
\arguments{
\item{basis}{The return value of function \code{CubicSmoothingSpline.Train}.}

\item{x_test}{The input values at which evaluations are required.}
}
\value{
The evaluated output at x_test.
}
\description{
This function takes the coefficients trained by \code{CubicSmoothingSpline.Train} and evaluate the output at x_test
}
\examples{
x<-seq(0, 1, 0.0015)
y <- x^3 * 3 - x^2 * 2 + x + exp(1)+rnorm(length(x),0,0.1)
plot(x,y)
lambda<-0.001
basis<-css_train(x,y,lambda)

x_test<-seq(0, 1, 0.1)
fit<-css_predict(basis,x_test)

plot(x_test,fit)
lines(x_test,x_test^3 * 3 - x_test^2 * 2 + x_test + exp(1),col="red")
}
