% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingspline.R
\name{css_train}
\alias{css_train}
\title{Train a smoothing spline with squared 2nd derivative penalty using natural cubic splines}
\usage{
css_train(x, y, lambda)
}
\arguments{
\item{x}{The input vector of training dataset.}

\item{y}{The output vector of training dataset.}

\item{lambda}{A fixed smoothing parameter.}
}
\value{
A list with the following components:
\item{beta}{ The coefficients of natural splines.}
\item{S}{The smoother matrix.}
\item{knots}{The knots used to construct the B-splines, including innerknots, boundary knots and phantom knots}
}
\description{
This function trains a smoothing spline with squared 2nd derivative penalty.
It has an explicit, finite-dimensional, 
unique minimizer which is a natural cubic spline.
}
\examples{
x<-seq(0, 1, 0.001)
y <- x^3 * 3 - x^2 * 2 + x + exp(1)+rnorm(length(x),0,0.1)
plot(x,y)
lambda<-0.001

basis<-css_train(x,y,lambda)
cat("the knots chosen are: ",basis$knots)
}
\references{
Friedman, J., Hastie, T., & Tibshirani, R. (2001). The Elements of Statistical Learning (Vol. 1, pp. 337-387). New York: Springer series in statistics.
 Chapter 5.4.
}
