% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicPowerBasisSpline.R
\name{CubicPowerBasisSpline}
\alias{CubicPowerBasisSpline}
\title{Regression using cubic spline}
\usage{
CubicPowerBasisSpline(x, y, x_test, innerknots)
}
\arguments{
\item{x}{The input vector of training dataset.}

\item{y}{The output vector of training dataset.}

\item{x_test}{The input values at which evaluations are required.}

\item{innerknots}{The internal knots that define the spline.}
}
\value{
A list with the following components:
\item{beta}{ The coefficients of nonparametric regression.}
\item{basis}{The cubic spline basis matrix of dimension c(length(x), NumKnots+4)}
\item{f}{The evaluated output at x_test.}
}
\description{
This function provides regressions using cubic splines. The cubic splines are defined following
Only univariate input can be used.
}
\examples{
n <- 100
t <- seq(0,2*pi,length.out = 100)
a <- 3
b <- 2
c.unif <- runif(n)
amp <- 2
set.seed(1)
y1 <- a*sin(b*t)+c.unif*amp # uniform error
innerknots <- 2*pi*c(1/4,2/4,3/4)
solution <- CubicPowerBasisSpline(t,y2,t,innerknots)
y.hat <- solution$f
plot(t, y1, t="l")
lines(t, y.hat, col=4)
}
