% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{eval_basis_functions}
\alias{eval_basis_functions}
\title{Evaluate basis functions as each x and return the evaluated basis matrix N}
\usage{
eval_basis_functions(x, knots, nknots)
}
\arguments{
\item{x}{Predictor variable vector}

\item{knots}{Knots location in terms of quantiles of x_train, optional, default will be evenly
spaced quantiles based on number of knots}

\item{nknots}{Number of knots useded in training.}
}
\value{
Basis matrix evaluated at each x value
}
\description{
Evaluate basis functions as each x and return the evaluated basis matrix N
}
