% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric_regression.R
\name{generate_knots}
\alias{generate_knots}
\title{Generate knots when real value is not specified.}
\usage{
generate_knots(x_train, df, knots, q)
}
\arguments{
\item{x_train}{The input vector of training dataset.}

\item{df}{Degrees of freedom. One can supply df rather than knots;
\code{generate_knots} then chooses (df + 1) knots at uniform quantiles of x.
The default, df = 4, sets 5 knots with 3 inner knots at uniform quantiles of x.}

\item{knots}{Breakpoints that define the spline, in terms of quantiles or real valus of x.
The default is five knots at uniform quantiles c(0, .25, .5, .75, 1).
Typical values are the mean or median for one knot, quantiles for more knots.}

\item{q}{A boolean variable define whether \code{knots} provided are quantiles or real values. When \code{q}=TRUE, \code{knots}
provided are quantiles of x. When \code{q}=FALSE, \code{knots} provided are real values of x.}
}
\value{
A vector of knots in terms of real values of x.
}
\description{
Generate knots when real value is not specified.
}
