% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{natural_cubic_splines.eval_basis}
\alias{natural_cubic_splines.eval_basis}
\title{Evaluate basis functions as each x and return the evaluated basis matrix N}
\usage{
natural_cubic_splines.eval_basis(x, knots, nknots)
}
\arguments{
\item{x}{Predictor variable vector.}

\item{knots}{Knots location in terms of quantiles of x_train, optional, default will be evenly
spaced quantiles based on number of knots.}

\item{nknots}{Number of knots useded in training.}
}
\value{
Basis matrix evaluated at each x value.
}
\description{
Evaluate basis functions as each x and return the evaluated basis matrix N
}
