% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natura_cubic_splines.R
\name{natural_cubic_splines.train}
\alias{natural_cubic_splines.train}
\title{Generate an evaluated basis matrix for natural cubic splines}
\usage{
natural_cubic_splines.train(x_train, y_train, df = NULL, knots = NULL,
  intercept = FALSE, Boundary.knots = range(x))
}
\arguments{
\item{Boundary.knots}{}
}
\value{
an evaluated basis matrix of size (length(x) * df)
}
\description{
Generate an evaluated basis matrix for natural cubic splines
}
