% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{ncs_basis}
\alias{ncs_basis}
\title{Generate an evaluated basis matrix for natural cubic splines}
\usage{
ncs_basis(x, knots)
}
\arguments{
\item{x}{Predictor variable vector.}

\item{knots}{Knots location in terms of real values of x.}
}
\value{
Basis matrix evaluated at each x value.
}
\description{
Generate an evaluated basis matrix for natural cubic splines
}
\examples{
x<-seq(0, 1, 0.001)
knots <- seq(0, 1, 0.1)

basis<-ncs_basis(x,knots)
plot(x,rep(0,length(x)),type="l",ylim=c(0,1))
for (i in 1: (length(knots))){
  lines(x,basis[,i],col=i)
}
}
