% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{ncs_predict}
\alias{ncs_predict}
\title{Prediction using regression spline with natural cubic spline.}
\usage{
ncs_predict(x_test, knots = NULL, beta = NULL, basis = NULL)
}
\arguments{
\item{x_test}{The input values at which evaluations are required.}

\item{knots}{Breakpoints that define the spline. \code{knots} should be in terms of real-values of x
It can be the return value of \code{generate_knots}.}

\item{beta}{The coefficients of nonparametric regression.}

\item{basis}{The return value of function \code{ncs_train}.
Instead of specify \code{knots} and \code{beta},One can supply \code{basis} directly.}
}
\value{
\item{y_pred}{A vector of dimension length(x), the prediction vector evaluated at x_test values.}
}
\description{
Prediction using regression spline with natural cubic spline.
}
