% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{ncs_train}
\alias{ncs_train}
\title{Generate an evaluated basis matrix for natural cubic splines}
\usage{
ncs_train(x_train, y_train, df = NULL, knots = NULL)
}
\arguments{
\item{x_train}{The input vector of training dataset.}

\item{y_train}{The output vector of training dataset.}

\item{df}{Degrees of freedom. One can supply df rather than knots;
ncs() then chooses (df + 1) knots at uniform quantiles of x.
The default, df = 4, sets 5 knots with 3 inner knots at uniform quantiles of x.}

\item{knots}{Breakpoints that define the spline, in terms of quantiles of x.
The default is five knots at uniform quantiles c(0, .25, .5, .75, 1).
Typical values are the mean or median for one knot, quantiles for more knots.}
}
\value{
A list of following components:
\item{nknots}{Number of knots.}
\item{knots}{A vector of knot locations.}
\item{N}{Basis matrix evaluated at each x value.}
\item{betas}{Least sqaure fit parameters.}
}
\description{
Generate an evaluated basis matrix for natural cubic splines
}
\examples{
x_train <- seq(1, 10, 0.1)
y_train <- cos(x_train)^3 * 3 - sin(x_train)^2 * 2 + x_train + exp(1)+rnorm(length(x_train),0,1)
plot(x_train,y_train)
x_test <- seq(1, 10, 0.1)
df <- 10
train_result <- ncs_train(x_train, y_train, df)
print(train_result$betas)
print(train_result$N[1:5,1:5])
}
