% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{ns}
\alias{ns}
\title{Generate a basis matrix for natural cubic splines}
\usage{
ns(x, df = NULL, knots = NULL, intercept = FALSE,
  Boundary.knots = range(x))
}
\arguments{
\item{Boundary.knots}{}
}
\value{
a basis matrix of size (length(x) * df)
}
\description{
Generate a basis matrix for natural cubic splines
}
