% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_basis_spline.R
\name{pbs_basis}
\alias{pbs_basis}
\title{Evaluate basis functions as each x and return the evaluated basis matrix N}
\usage{
pbs_basis(x, order, knots)
}
\arguments{
\item{x}{Predictor variable vector.}

\item{knots}{Knots location in terms of quantiles of x_train, optional, default will be evenly
spaced quantiles based on number of knots.}

\item{nknots}{Number of knots useded in training.}
}
\value{
Basis matrix evaluated at each x value.
}
\description{
Evaluate basis functions as each x and return the evaluated basis matrix N
}
