% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric_regression.R
\name{place_knots}
\alias{place_knots}
\title{Find evenly spaced knots by quantile}
\usage{
place_knots(nknots, x)
}
\arguments{
\item{nknots}{Number of knots to be located.}

\item{x}{Data vector on which knots are placed.}
}
\value{
A named vector with knot quantiles and values.
}
\description{
Knots found include boundary knots at 0th and 100th quantile.
}
