% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerBasisSpline.R
\name{PowerBasisSpline}
\alias{PowerBasisSpline}
\title{Regression using Power Basis spline}
\usage{
PowerBasisSpline(x, y, x_test, order, innerknots)
}
\arguments{
\item{x}{The input vector of training dataset.}

\item{y}{The output vector of training dataset.}

\item{x_test}{The input values at which evaluations are required.}

\item{order}{The order that defines the spline.}

\item{innerknots}{The internal knots that define the spline.}
}
\value{
A list with the following components:
\item{beta}{ The coefficients of nonparametric regression.}
\item{basis}{The spline basis matrix of dimension c(length(x), NumKnots+order)}
\item{f}{The evaluated output at x_test.}
}
\description{
This function is a generalization of CubicPowerBasisSpline with arbitrary order
}
\details{
Only univariate input can be used.
}
\examples{
n <- 100
t <- seq(0,2*pi,length.out = 100)
a <- 3
b <- 2
c.unif <- runif(n)
amp <- 2
set.seed(1)
y1 <- a*sin(b*t)+c.unif*amp # uniform error
innerknots <- 2*pi*c(1/4,2/4,3/4)
order <- 4
solution <- PowerBasisSpline(t,y1,t,order,innerknots)
y.hat <- solution$f
plot(t, y1, t="l")
lines(t, y.hat, col=2)
}
