% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R, R/natural_cubic_splines.R
\name{sel_smoothing_para}
\alias{sel_smoothing_para}
\alias{sel_smoothing_para}
\title{Select smoothing parameter based on leave-one-out CV error}
\usage{
sel_smoothing_para(x, y, cv_lambda)

sel_smoothing_para(x, y, cv_lambda)
}
\arguments{
\item{x}{predictor variable}

\item{y}{response variable}

\item{cv_lambda}{vector of candidate lambda values}

\item{x}{predictor variable.}

\item{y}{response variable.}

\item{cv_lambda}{vector of candidate lambda values, must be between 0 and 1.}
}
\value{
lamdba value that minimizes leave-one-out CV error

lamdba value that minimizes leave-one-out CV error.
}
\description{
Select smoothing parameter based on leave-one-out CV error

Select smoothing parameter based on leave-one-out CV error
}
