% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{basis_function}
\alias{basis_function}
\title{Evalute x based on truncated power basis functions for natural cubic splines}
\usage{
basis_function(x, i)
}
\arguments{
\item{i, i}{= 1, ..., length(x)}
}
\description{
Evalute x based on truncated power basis functions for natural cubic splines
}
