% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bs_predict}
\alias{bs_predict}
\title{Prediction using regression spline with B-spline basis}
\usage{
bs_predict(x_test, order = NULL, knots = NULL, beta = NULL,
  basis = NULL)
}
\arguments{
\item{x_test}{The input values at which evaluations are required.}

\item{order}{The order of basis functions. order=degree+1}

\item{knots}{Breakpoints that define the spline. \code{knots} should be in terms of real-values of x
and contain innner, boundary and phantom knots. It can be the return value of \code{bs_knots}.}

\item{beta}{The coefficients of nonparametric regression.}

\item{basis}{The return value of function \code{bs_train}.
Instead of specify \code{knots}, \code{order} and \code{beta},One can supply \code{basis} directly.}
}
\value{
The evaluated output at x_test.
}
\description{
This function provides prediction at value of interest using regression spline with B-spline basis.
The B-splines are generated by \code{bs_basis} and trained by the \code{bs_train}.
The return value of \code{bs_train} can be used as an argument of \code{bs_predict}
}
\examples{
x<-seq(0, 1, 0.001)
y <- x^3 * 3 - x^2 * 2 + x + exp(1)+rnorm(length(x),0,0.1)
plot(x,y)
knots <- seq(0.1, 0.9, 0.01)
order<-4
basis<-bs_train(x,y,order,knots)

x_test<-seq(0, 1, 0.01)
fit<-bs_predict(x_test,basis=basis)
plot(x_test,fit)
lines(x_test,x_test^3 * 3 - x_test^2 * 2 + x_test + exp(1),col="red")
}
\seealso{
\code{bs_basis},\code{bs_knots}, \code{bs_train}, \code{np_reg}.
}
