% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bsplineBasis}
\alias{bsplineBasis}
\title{Generating B-spline basis}
\usage{
bsplineBasis(x, y, order, innerknots)
}
\arguments{
\item{x}{The input vector of training dataset.}

\item{y}{The output vector of training dataset.}

\item{order}{The order of B-spline functions. The default is order=4 for cubic B-splines.}

\item{innerknots}{The internal knots that define the spline. \code{innerknots} should not contain knots on the boundary.}
}
\value{
A list with the following components:
\item{beta}{ The coefficients of nonparametric regression.}
\item{basis}{The B-spline basis matrix of dimension c(length(x), df). df = length(innerknots) + order.}
\item{knots}{The knots used to construct the B-splines, including innerknots, boundary knots and phantom knots}
\item{order}{The order of basis functions. order=degree+1}
}
\description{
This function generates B-spline basis. The B-splines are defined following
the recursive formulas due to de Boor. Only univariate input can be used.
}
\examples{
x<-seq(0, 1, 0.001)
y <- x^3 * 3 - x^2 * 2 + x + exp(1)+rnorm(length(x),0,0.1)
plot(x,y)
innerknots <- seq(0.1, 0.9, 0.1)
order<-4

basis<-bsplineBasis(x,y,order,innerknots)
plot(x,rep(0,length(x)),type="l",ylim=c(0,1))
for (i in 1: (length(innerknots)+order)){
  lines(x,basis$basismatrix[,i])
}
}
