% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natura_cubic_splines_test.R
\name{natural_cubic_splines.predict}
\alias{natural_cubic_splines.predict}
\title{Prediction based on trained regression model}
\usage{
natural_cubic_splines.predict(x_test)
}
\arguments{
\item{x_test}{}
}
\description{
Prediction based on trained regression model
}
