% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{natural_cubic_splines}
\alias{natural_cubic_splines}
\title{Regression using natural cubic splines}
\usage{
natural_cubic_splines(x_train, y_train, x_test, df = NULL, knots = NULL)
}
\arguments{
\item{x_train}{The input vector of training dataset.}

\item{y_train}{The output vector of training dataset.}

\item{x_test}{The input values at which evaluations are required.}

\item{df}{The degree of freedom specified by user, number of knots will be equal to df.}

\item{knots}{Knots location in terms of quantiles of x_train, optional, default will be evenly
spaced quantiles based on number of knots}
}
\value{
\item{y_pred}{A vector of dimension length(x)The prediction vector evaluated at x_test values}
}
\description{
This function provides regressions using natural cubic splines with truncated power basis functions.
Only univariate input can be used.
}
\examples{
x_train <-seq(0, 1, 0.001)
y_train <- x^3 * 3 - x^2 * 2 + x + exp(1)+rnorm(length(x),0,0.1)
plot(x,y)
df <- 10
x_test <- seq(0, 1, 0.01)
y_pred <- natural_cubic_splines(x, y, x_test, df)
plot(x_test,y_pred)
lines(x_test,x_test^3 * 3 - x_test^2 * 2 + x_test + exp(1),col="red")
}
