% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_cubic_splines.R
\name{ncs_train}
\alias{ncs_train}
\title{Train regression coefficients for natural cubic splines.}
\usage{
ncs_train(x_train, y_train, df = NULL, knots = NULL, q = FALSE)
}
\arguments{
\item{x_train}{The input vector of training dataset.}

\item{y_train}{The output vector of training dataset.}

\item{df}{Degrees of freedom. One can supply df rather than knots;
\code{ncs_train} then chooses (df + 1) knots at uniform quantiles of x.
The default, df = 4, sets 5 knots with 3 inner knots at uniform quantiles of x.}

\item{knots}{Breakpoints that define the spline, in terms of quantiles of x or real values of x.
The default is five knots at uniform quantiles c(0, .25, .5, .75, 1).
Typical values are the mean or median for one knot, quantiles for more knots.}

\item{q}{A boolean variable indicating whether \code{knots} provided are quantiles or real values. When \code{q}=TRUE, \code{knots}
provided are quantiles of x. When \code{q}=FALSE, \code{knots} provided are real values of x. Default is FALSE.}
}
\value{
A list of following components:
\item{nknots}{Number of knots.}
\item{knots}{A vector of knot locations.}
\item{N}{Basis matrix evaluated at each x value.}
\item{betas}{Least sqaure fit parameters.}
}
\description{
In the least square fitting of nonparametric egression coefficients,
Moore-Penrose generalized inverse (ginv\{MASS\}) is used to aviod computational problems.
}
\examples{
x_train <- seq(1, 10, 0.1)
y_train <- cos(x_train)^3 * 3 - sin(x_train)^2 * 2 + x_train + exp(1)+rnorm(length(x_train),0,1)
plot(x_train,y_train)
x_test <- seq(1, 10, 0.1)
df <- 10
train_result <- ncs_train(x_train, y_train, df)
print(train_result$beta)
print(train_result$N[1:5,1:5])
}
\references{
Friedman, J., Hastie, T., & Tibshirani, R. (2001). The Elements of Statistical Learning (Vol. 1, pp. 337-387). New York: Springer series in statistics.
Chapter 5.2.1.

Venables, W. N. and Ripley, B. D. (1999) Modern Applied Statistics with S-PLUS. Third Edition. Springer. p.100.
}
