% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_basis_spline.R
\name{pbs_train}
\alias{pbs_train}
\title{Regression using truncated power basis spline}
\usage{
pbs_train(x, y, order, df = NULL, knots = NULL, q = FALSE)
}
\arguments{
\item{x}{The input vector of training dataset.}

\item{y}{The output vector of training dataset.}

\item{order}{The order that defines the spline.}

\item{df}{Degrees of freedom. One can supply df rather than knots.}

\item{knots}{The innerknots and boundary knots that define the spline.
The knots provided can be quantiles of x or real values of x.
More explanation of \code{knots}, \code{df}, \code{q} can be seen in \code{generate_knots}.}

\item{q}{A boolean variable define whether \code{knots} provided are quantiles or real values. When \code{q}=TRUE, \code{knots}
provided are quantiles of x. When \code{q}=FALSE, \code{knots} provided are real values of x. Default is FALSE.}

\item{x_test}{The input values at which evaluations are required.}
}
\value{
A list with the following components:
\item{beta}{ The coefficients of nonparametric regression.}
\item{basis}{The spline basis matrix of dimension c(length(x), length(knots)+order)}
\item{knots}{The knots used to construct the power basis splines}
\item{order}{The order of basis functions. order=degree+1}
}
\description{
This function provides regressions using truncated power basis spline. The basis are defined as
1,x,x^2,...,x^m,(x-k1)^(m-1)+,(x-k2)^(m-1)+,...,(x-kn)^(m-1)+
where m is the order, k1, k2 and kn are n knots, '+' denotes the positive part.
}
\details{
Only univariate input can be used.
}
\examples{
n <- 100
t <- seq(0,2*pi,length.out = 100)
a <- 3
b <- 2
c.unif <- runif(n)
amp <- 2
set.seed(1)
y1 <- a*sin(b*t)+c.unif*amp # uniform error
knots <- c(min(t),2*pi*c(1/4,2/4,3/4),max(t))
order <- 4
basis <- pbs_train(t,y1,order,knots=knots)
cat("trained coeffecients for every spline are",basis$beta)
}
\references{
Friedman, J., Hastie, T., & Tibshirani, R. (2001). The Elements of Statistical Learning (Vol. 1, pp. 337-387). New York: Springer series in statistics.
Chapter 5.2.1.
}
\seealso{
\code{generate_knots}.
}
